backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
tortoise-orm>=0.18.0
aiomysql>=0.0.22
pydantic>=1.9.0
python-dotenv>=0.19.0
aerich>=0.6.3  # Para migraciones
